CI: use a copy of master project but locked
This is for PR testing. In order to prevent issues when packages are
being rebuild, which happens for example when packages are in a state A
when deploying, and state B when running the tests. Some tests will try
to install packages.

Having an error because of inconsistent metadata at the test it is very
paintful because we have to retrigger a 2 hour job.

Instead, we have a project for each environment which has links to all
packages in sm:Master. This way, we can disable the build before
running the tests.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>