Use lockfile for concurrent PR e2e testing
We were using test and touch for "acquiring" the test environment.

However, during our tests, we found out that two jobs could, and
actually did, take the same environment.

Most probably, this happened because two jobs run the "test" before one
of them could do the "touch".

Instead, let's use "lockfile" command, which implements semaphores,
which is what actually we were trying to implement with the "test and
touch".

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>