Drop parallel_client tests experimental feature
This feature can't work because we cannot connect to the web ui with the
same user at the same time. They will share the session and the
selection in the UI.

We could fix this by using different users for different clients, but
this is a more complex thing to do, so for now we drop this feature.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>