Skip openSUSE mirrors for testing Pull Requests
Downloading recently build packages from openSUSE mirrors is error
prone. Packages can simply not yet be updated or available, out of sync,
etc.. metadata can be outdated... not to talk about the fact that
sometimes the build service publisher job gets stucked and we do not get
the package published for hours.

Instead, we are going to download packages from the build service
directly into the jenkins worker. Then, we will create the metadata and
"expose" those repos via a web server.

We were already doing this for the packages build from the Pull Request.
We are extending the same idea for all repos.

This has an advantage: we do not need anymore to have a project in the
build service for each environment, nor we need to lock and unlock it.
Downloading the binaries into a different directory for each environment
would do the trick, instead.

Given we check if packages build in the devel repo, and not in the CR
subprojects, we need to fail if packages fail to build on the target repo.

Otherwise, if we are testing for openSUSE and Almalinux fails, we cannot
test.

These was not needed before because the CR projects did not had
Almalinux or any repo which is not being tested.

In order for this to work, we need to disable the devel master branch
projects, and set some additional projects, instead.

For this reason, we have created a new product definition, uyuni-pr.

use official opensuse image for clients

Because we do not need to download all the packages each time, only the
new ones, we are keeping the downloaded packages. Except for the
packages built from the Pull Request. In that case, we need to remove
the old ones because we can not ensure the new ones will have a lower
release number. This is like this because we are recrating the project
on each built. Thus, a new build could have a lower number than a
previous built.

With this commit, we do not rely anymore on the openSUSE mirroring
infrastructure. This add stability to testing Pull Requests.