Use official image for testing clients in Pull Request
Using a custom image is not working. The image we built in obs has
network issues with libvirt.

Let's use the official one. In any case we don't donwload the image
every time, only if we do a "cleanup", which happens seldomly and is
done manually. Cross your fingers the image is not unavailable at that
moment.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>