Review email links for Pull Requests
Fix email links for pull request acceptance tests

The workspace is not a constant, as in the master tests, so we cannot
use a link to the workspace in the email.

Instead, let's archive what we need as artifacts and add links to that.

Send email even when there has been an error during the deploy phase

Check build result dir exists

Add some help links to emails

Add link to blue ocean pipelines