Review parameters: obs and sumaform
The user do not change the obs parameters, so let's make them variables
instead and remove them from the parameters list. This makes the UI
simpler and more understandable. We were using these parameters when
developing this feature, but not anymore.

Add some clarification on when to use the sumaform parameters.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>