Remove if not deployed
Makes no sense to keep the environment locked for an extra hour if we
could not deployed. For example, this happens if there is any error
prior to the deploy phase. In that case, several retries which will take
less than an hour, will end up with all environment locked.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>