Do not use openSUSE official repos
We are having trouble with the official repos. From time to time there
is an error. Thus, let's not use them at all.

Use minima mirror instead of download.opensuse.org

This is to minimize depending on external infrastructure such as
openSUSE mirrors.

Use an openSUSE image with no repositories at all for all openSUSE vms.

Add the repos with the tf_files, instead of using sumaform, so we can
add the ones in the minima mirror.

We are skipping update repos to avoid the problem with wrong metadata or
missing rpms. This way, it won't happen anymore plus it is
reproduceable.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>