Do not let user choose the sumaform repo. Simplify.
You can setup the sumaform via the environment file, which uses one or
the other value depending on the job name.

This way, we can still use a different sumaform, but it needs to be done
via changing that file. This is preferable because when changing the
sumaform, you need to do a manual clean up, so better do not make it too
easy for anyone to change this.