Let's not merge with git master
I believe merging with master makes the git fetch/clone take longer
because a shallow copy is not enough if we want to merge the PR with
master.

This has the disadvantage that we are not including latest fixes from
master, so the developer will have to rebase her branch.

It has the advantage of being more stable and predictable.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>