Use minima mirror for PR testing
We need this to be able to skip opensuse mirrors even before the
cloudinit. Otherwise, we get this error that salt-call is not available
in CentOS because there was an error installing salt from openSUSE
mirrors.

Also, if we do not do this, we were getting an error preparing the test
repos with the dummy  package while downloading the packages from
openSUSE mirror.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>