Clean up published RPMS for PR
RPMs for PR need to be recreated each time in the jenkins worker to
server as a "mirror" for the tests.

If we do not clean up them at the end of the pipeline, they are left in
the disc space as trash.

We run out of disc space because of this.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>