Add pipeline for testing PRs with SUMA 4.3
1- Refactor: extract differences with uyuni and suma43 as variables

This is the first step to generalize the pipeline, so that it can be
used for both, uyuni and suma43.

Note, we need to create the dirs /storage/repos/suma/[1-10]
/storage/repos/suma43/[5-6]

2- Add suma43 pipeline

Extract the variables to the env.groovy file and add the suma43 job.

3- Define terraform conf files for SUMA 4.3 Acceptance tests PR testing

Based on what we have for Uyuni. Let's use hyperion.