Remove unneeded images
We do not need anymore our own leap 5.5 custom image. That image was
used for preinstalling packages for the server and the proxy. Given we
use the leap micro version for those images, our custom image can be
removed.

Instead, the default is the opensuse 5.5 . This image is used for
kvm-host.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>