Move acceptance tests from PRs to NUE
Running this jenkins pipeline at the same place we have the
open build service should make the process faster and less complex.

However, in NUE we have less hardware, so we can have max 6 envs.

Given we have only 6 envs in NUE and not 10 as we have in PRV, we need
to do a cleanup, specifically remove the "custom" pipelines, like
"jordi", "qe-squad", and "iron" ones.

This way, we have 3 for generic master PR testing, 1 for 4.3 PR, one for
master reference, and one for 4.3 reference.

Fix rhlike and deblike minion names for PR Testing in NUE

Fix domain for pr testing: domain should depend on the localization

fix deblike macs in NUE

Share environments between uyuni and suma4.3, because we have less
hardware but also less users, thus we do not need to rebase.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>