CI: acceptance tests at PR: cleanup environments
Given we have only 6 envs in NUE and not 10 as we have in PRV, we need
to do a cleanup, specifically remove the "custom" pipelines, like
"jordi", "qe-squad", and "iron" ones.

This way, we have 3 for generic master PR testing, 1 for 4.3 PR, one for
master reference, and one for 4.3 reference.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>